correction to nyears for totbio
something odd occurs where the terminal group of projected quantities (corresponding to tot_biom_proj) are of variable lengths; sometimes only 13 years, sometimes 16.  Unclear where the error is occurring since a) the .tpl isn't changing and b) the .std files are of uniform length (15 years).

This leverages the fact they always occur at the end to manually assign years and simply pastes the year labels so the overall csv can be compiled and written; total biomass is not included in the plot.